fix: use c++ memory allocator to match delete calls
Use ::operator new(size) to allocate memory inside C++ instead of using
the Rust libc::malloc. This as the Rust call maps to the C malloc
function rather than the C++ new keyword. As delete keyword is used to
free memory, we also must allocate with new to ensure memory is correctly
managed.